requirements.txt
flask
gunicorn  # WSGI server for production
pandas
numpy
plotly
pyyaml
scikit-learn
scipy
flask-socketio